Card component rectified
Inline stylings removed
