+++ package.json
@@ -12,7 +12,7 @@
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.0",
     "uuid": "^8.3.2"
   },
   "scripts": {
