+++ src/components/authentication/LogInForm.jsx
@@ -17,8 +17,6 @@ function LogInForm() {
     e.preventDefault();
 
     logInHandler(details.email, details.password);
-
-    console.log("triggered");
   }
 
   return (
